requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic>=2,<3
python-dateutil>=2.9,<3
httpx>=0.27,<1
pytest>=8,<9
